Keep text inside triple backquotes intact
